requirements.txt
fastapi
jinja2
python-dotenv
python-multipart
requests
supabase
uvicorn
